backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
gunicorn==21.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
sqlalchemy>=2.0.29
pydantic>=2.6.1
pydantic-settings>=2.1.0
python-dotenv==1.0.1
ccxt==4.2.15
websockets==12.0
bcrypt==4.1.2 